Create reduction_potential_dataset
This dataset presents experimental reduction potential values of Pt(IV) complexes collected from the literature. Reduction potentials are expressed in V and for each paper refer to a precise electrode. For conversion of the values in reference to the standard hydrogen electrode (SHE) consult the methods sections of the referenced article: https://doi.org/10.1021/acs.jcim.4c00315.
If you want to use this dataset please cite: J. Chem. Inf. Model. 2024, 64, 9, 3733-3743.